csv-to-score: abrupt return to baseline
Make sure that a decline at the beginning of the night gets picked up,
but one at the end of the night does not. Enumerate baselines from 0.
Write `abrupts` in a more readable way.
Ensure pressures::[Double] and references::[Double] don't get mixed up.
Test that abrupts correctly rejects a decline it accepted in the past.