NoxPes2Csv: replace swallows and coughs with zeros.
Exactly zero voltage from the pressure sensor is improbable,
except as an imputed value. We use positive zero as an
indication of a missing pressure measurement or non-breathing activity.

csv-to-score is expected to later know to ignore these, or better yet,
use them as barriers across which Pes crescendos can not extend.