overlap/agreementTotals.svg: a bar chart of a lack of a consensus
There is a lot of sleep that all methods can agree on being absolutely normal.
The top bar is for sleep that matches 0 of the studied respiratory effort patterns.
There is also quite a lot of sleep that matches 1 pattern but no other.
Equivalently, there is also a lot of sleep that all but one pattern agree is normal.
There is half as much sleep that all but two pattern agree is normal.
There is less sleep that all but three patterns agree is normal, and so on,
and in the end we find that there is no sleep that no pattern considers normal.

That is, all sleep is considered normal by some pattern.